requirements.txt
aniso8601==10.0.0
attrs==25.1.0
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
clams==0.0.5
clams-python==1.3.1
click==8.1.8
deepdiff==8.2.0
ffmpeg-python==0.2.0
Flask==3.0.3
Flask-RESTful==0.3.10
future==1.0.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lapps==0.0.2
lxml==5.3.0
MarkupSafe==2.1.5
mmif-python==1.0.19
numpy==1.24.3
opencv-python==4.11.0.86
orderly-set==5.3.0
packaging==24.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pydantic==1.10.21
pytesseract==0.3.10
pytz==2025.1
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.20.1
six==1.17.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
zeep==4.3.1
zipp==3.20.2
